requirements.txt
openai>=1.0.0
feedparser>=6.0.0
Pillow>=10.0.0
requests>=2.28.0
python-dateutil>=2.8.0
